Added Object Type property

+++ D2DPlatformer/Background.cpp
@@ -3,7 +3,7 @@
 
 Background::~Background()
 {
-	for (int i = 0; i < m_layers.size(); i++)
+	for (size_t i = 0; i < m_layers.size(); i++)
 		delete m_layers[i];
 }
